+++ pdata/internal/data/spanid_test.go
@@ -20,13 +20,13 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNewSpanID(t *testing.T) {
-	sid := NewSpanID([8]byte{})
+func TestSpanID(t *testing.T) {
+	sid := SpanID([8]byte{})
 	assert.EqualValues(t, [8]byte{}, sid)
 	assert.EqualValues(t, 0, sid.Size())
 
 	b := [8]byte{1, 2, 3, 4, 5, 6, 7, 8}
-	sid = NewSpanID(b)
+	sid = b
 	assert.EqualValues(t, b, sid)
 	assert.EqualValues(t, 8, sid.Size())
 }
@@ -34,12 +34,12 @@ func TestNewSpanID(t *testing.T) {
 func TestSpanIDMarshal(t *testing.T) {
 	buf := make([]byte, 10)
 
-	sid := NewSpanID([8]byte{})
+	sid := SpanID([8]byte{})
 	n, err := sid.MarshalTo(buf)
 	assert.EqualValues(t, 0, n)
 	assert.NoError(t, err)
 
-	sid = NewSpanID([8]byte{1, 2, 3, 4, 5, 6, 7, 8})
+	sid = [8]byte{1, 2, 3, 4, 5, 6, 7, 8}
 	n, err = sid.MarshalTo(buf)
 	assert.NoError(t, err)
 	assert.EqualValues(t, 8, n)
@@ -50,12 +50,12 @@ func TestSpanIDMarshal(t *testing.T) {
 }
 
 func TestSpanIDMarshalJSON(t *testing.T) {
-	sid := NewSpanID([8]byte{})
+	sid := SpanID([8]byte{})
 	json, err := sid.MarshalJSON()
 	assert.EqualValues(t, []byte(`""`), json)
 	assert.NoError(t, err)
 
-	sid = NewSpanID([8]byte{0x12, 0x23, 0xAD, 0x12, 0x23, 0xAD, 0x12, 0x23})
+	sid = [8]byte{0x12, 0x23, 0xAD, 0x12, 0x23, 0xAD, 0x12, 0x23}
 	json, err = sid.MarshalJSON()
 	assert.EqualValues(t, []byte(`"1223ad1223ad1223"`), json)
 	assert.NoError(t, err)