+++ pdata/internal/data/traceid_test.go
@@ -20,13 +20,13 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNewTraceID(t *testing.T) {
-	tid := NewTraceID([16]byte{})
+func TestTraceID(t *testing.T) {
+	tid := TraceID([16]byte{})
 	assert.EqualValues(t, [16]byte{}, tid)
 	assert.EqualValues(t, 0, tid.Size())
 
 	b := [16]byte{0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78}
-	tid = NewTraceID(b)
+	tid = b
 	assert.EqualValues(t, b, tid)
 	assert.EqualValues(t, 16, tid.Size())
 }
@@ -34,12 +34,12 @@ func TestNewTraceID(t *testing.T) {
 func TestTraceIDMarshal(t *testing.T) {
 	buf := make([]byte, 20)
 
-	tid := NewTraceID([16]byte{})
+	tid := TraceID([16]byte{})
 	n, err := tid.MarshalTo(buf)
 	assert.EqualValues(t, 0, n)
 	assert.NoError(t, err)
 
-	tid = NewTraceID([16]byte{0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78})
+	tid = [16]byte{0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78}
 	n, err = tid.MarshalTo(buf)
 	assert.EqualValues(t, 16, n)
 	assert.EqualValues(t, []byte{0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78}, buf[0:16])
@@ -50,12 +50,12 @@ func TestTraceIDMarshal(t *testing.T) {
 }
 
 func TestTraceIDMarshalJSON(t *testing.T) {
-	tid := NewTraceID([16]byte{})
+	tid := TraceID([16]byte{})
 	json, err := tid.MarshalJSON()
 	assert.EqualValues(t, []byte(`""`), json)
 	assert.NoError(t, err)
 
-	tid = NewTraceID([16]byte{0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78})
+	tid = [16]byte{0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78, 0x12, 0x34, 0x56, 0x78}
 	json, err = tid.MarshalJSON()
 	assert.EqualValues(t, []byte(`"12345678123456781234567812345678"`), json)
 	assert.NoError(t, err)
@@ -79,7 +79,7 @@ func TestTraceIDUnmarshal(t *testing.T) {
 }
 
 func TestTraceIDUnmarshalJSON(t *testing.T) {
-	tid := NewTraceID([16]byte{})
+	tid := TraceID([16]byte{})
 	err := tid.UnmarshalJSON([]byte(`""`))
 	assert.NoError(t, err)
 	assert.EqualValues(t, [16]byte{}, tid)