+++ pdata/pmetric/generated_metrics_test.go
@@ -1063,18 +1063,16 @@ func TestHistogramDataPoint_Sum(t *testing.T) {
 
 func TestHistogramDataPoint_BucketCounts(t *testing.T) {
 	ms := NewHistogramDataPoint()
-	assert.Equal(t, pcommon.NewImmutableUInt64Slice([]uint64(nil)), ms.BucketCounts())
-	testValBucketCounts := pcommon.NewImmutableUInt64Slice([]uint64{1, 2, 3})
-	ms.SetBucketCounts(testValBucketCounts)
-	assert.Equal(t, testValBucketCounts, ms.BucketCounts())
+	assert.Equal(t, []uint64(nil), ms.BucketCounts().AsRaw())
+	ms.BucketCounts().FromRaw([]uint64{1, 2, 3})
+	assert.Equal(t, []uint64{1, 2, 3}, ms.BucketCounts().AsRaw())
 }
 
 func TestHistogramDataPoint_ExplicitBounds(t *testing.T) {
 	ms := NewHistogramDataPoint()
-	assert.Equal(t, pcommon.NewImmutableFloat64Slice([]float64(nil)), ms.ExplicitBounds())
-	testValExplicitBounds := pcommon.NewImmutableFloat64Slice([]float64{1, 2, 3})
-	ms.SetExplicitBounds(testValExplicitBounds)
-	assert.Equal(t, testValExplicitBounds, ms.ExplicitBounds())
+	assert.Equal(t, []float64(nil), ms.ExplicitBounds().AsRaw())
+	ms.ExplicitBounds().FromRaw([]float64{1, 2, 3})
+	assert.Equal(t, []float64{1, 2, 3}, ms.ExplicitBounds().AsRaw())
 }
 
 func TestHistogramDataPoint_Exemplars(t *testing.T) {
@@ -1353,10 +1351,9 @@ func TestBuckets_Offset(t *testing.T) {
 
 func TestBuckets_BucketCounts(t *testing.T) {
 	ms := NewBuckets()
-	assert.Equal(t, pcommon.NewImmutableUInt64Slice([]uint64(nil)), ms.BucketCounts())
-	testValBucketCounts := pcommon.NewImmutableUInt64Slice([]uint64{1, 2, 3})
-	ms.SetBucketCounts(testValBucketCounts)
-	assert.Equal(t, testValBucketCounts, ms.BucketCounts())
+	assert.Equal(t, []uint64(nil), ms.BucketCounts().AsRaw())
+	ms.BucketCounts().FromRaw([]uint64{1, 2, 3})
+	assert.Equal(t, []uint64{1, 2, 3}, ms.BucketCounts().AsRaw())
 }
 
 func TestSummaryDataPointSlice(t *testing.T) {
@@ -1838,16 +1835,16 @@ func TestExemplar_FilteredAttributes(t *testing.T) {
 
 func TestExemplar_TraceID(t *testing.T) {
 	ms := NewExemplar()
-	assert.Equal(t, pcommon.TraceID(data.NewTraceID([16]byte{})), ms.TraceID())
-	testValTraceID := pcommon.TraceID(data.NewTraceID([16]byte{1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 6, 5, 4, 3, 2, 1}))
+	assert.Equal(t, pcommon.TraceID(data.TraceID([16]byte{})), ms.TraceID())
+	testValTraceID := pcommon.TraceID(data.TraceID([16]byte{1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 6, 5, 4, 3, 2, 1}))
 	ms.SetTraceID(testValTraceID)
 	assert.Equal(t, testValTraceID, ms.TraceID())
 }
 
 func TestExemplar_SpanID(t *testing.T) {
 	ms := NewExemplar()
-	assert.Equal(t, pcommon.SpanID(data.NewSpanID([8]byte{})), ms.SpanID())
-	testValSpanID := pcommon.SpanID(data.NewSpanID([8]byte{8, 7, 6, 5, 4, 3, 2, 1}))
+	assert.Equal(t, pcommon.SpanID(data.SpanID([8]byte{})), ms.SpanID())
+	testValSpanID := pcommon.SpanID(data.SpanID([8]byte{8, 7, 6, 5, 4, 3, 2, 1}))
 	ms.SetSpanID(testValSpanID)
 	assert.Equal(t, testValSpanID, ms.SpanID())
 }